+++ CliHelper.Tests/ParameterWithValuesTests.cs
@@ -1,0 +1,9 @@
+using NUnit.Framework;
+
+namespace CliHelper.Tests
+{
+    public class ParameterWithValuesTests
+    {
+        // TODO: GetValue Tests.
+    }
+}